Insert script completed
The script now inserts the header and footer from a template document into a html document. This only works when the html document only contains a body currently.

Readme also expanded and other changes are just documents used in testing; test.html being the test document and backup.html just being a backup of the test.html to accelerate testing.