Timestamp Script
Created a script to add/update a timestamp to the footer of a html document to show when it was last edited (Date). ID of the html element that contains the timestamp is "timestamp".

Updated insert script to run timestamp at the end of the script so to automatically update.