Ranking Script completed
Finished ranking script so it updates the university's rank in a span. Could be tweaked so that in the edge case that another span exists with a number in it, that won't also be modified. For now do not have a span with just a number in it.

Also added timestamp to be run on the html document when insert or ranking is run to automatically update the last edit timestamp.